# packages/contracts/openapi.json
{
 "components": {
  "schemas": {
...
   "BulkDecision": {
    "properties": {
     "decision": {
      "enum": [
       "accept",
       "reject",
       "defer"
      ],
      "title": "Decision",
      "type": "string"
     },
     "ids": {
      "items": {
       "format": "uuid",
       "type": "string"
      },
      "title": "Ids",
      "type": "array"
     },
     "rationale": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rationale"
     }
    },
    "required": [
     "ids",
     "decision"
    ],
    "title": "BulkDecision",
    "type": "object"
   },
   "ConversationIn": {
    "properties": {
     "curriculum_version_id": {
      "format": "uuid",
      "title": "Curriculum Version Id",
      "type": "string"
     },
     "scenario_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Scenario Id"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "required": [
     "curriculum_version_id"
    ],
    "title": "ConversationIn",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "MessageIn": {
    "properties": {
     "mode": {
      "default": "explore",
      "enum": [
       "explore",
       "investigate",
       "simulate"
      ],
      "title": "Mode",
      "type": "string"
     },
     "text": {
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "MessageIn",
    "type": "object"
   },
...
   "ProgramIn": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "institution": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Institution"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "parent_degree": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Degree"
     }
    },
    "required": [
     "code",
     "name"
    ],
    "title": "ProgramIn",
    "type": "object"
   },
   "ProgramOut": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "institution": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Institution"
     },
     "is_synthetic": {
      "title": "Is Synthetic",
      "type": "boolean"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "parent_degree": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Degree"
     }
    },
    "required": [
     "id",
     "code",
     "name",
     "institution",
     "parent_degree",
     "description",
     "is_synthetic"
    ],
    "title": "ProgramOut",
    "type": "object"
   },
   "ProviderPolicy": {
    "properties": {
     "ai_providers": {
      "anyOf": [
       {
        "items": {
         "enum": [
          "openai",
          "anthropic"
         ],
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ai Providers"
     }
    },
    "required": [
     "ai_providers"
    ],
    "title": "ProviderPolicy",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Curriculum decision support. Not a predictor of learning gains or an accreditation certifier.",
  "title": "Curriculum Flight Simulator API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/ai/runs/{run_id}": {
   "get": {
    "operationId": "model_run_api_v1_ai_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Model Run",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/ai/status": {
   "get": {
    "operationId": "ai_status_api_v1_ai_status_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Ai Status Api V1 Ai Status Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ai Status",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/ai/usage": {
   "get": {
    "operationId": "ai_usage_api_v1_ai_usage_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Ai Usage Api V1 Ai Usage Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ai Usage",
    "tags": [
     "ai"
    ]
   }
  },
...
  "/api/v1/analyses/{run_id}/explain": {
   "post": {
    "operationId": "explain_api_v1_analyses__run_id__explain_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Explain",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/conversations": {
   "get": {
    "operationId": "list_conversations_api_v1_conversations_get",
    "parameters": [
     {
      "in": "query",
      "name": "curriculum_version_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "uuid",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Curriculum Version Id"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Conversations",
    "tags": [
     "ai"
    ]
   },
   "post": {
    "operationId": "create_conversation_api_v1_conversations_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConversationIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Conversation",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/conversations/{cid}/messages": {
   "get": {
    "operationId": "list_messages_api_v1_conversations__cid__messages_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Messages",
    "tags": [
     "ai"
    ]
   },
   "post": {
    "description": "Queues an assistant turn. Stream progress from /jobs/{job_id}/stream; the answer is stored as a message.",
    "operationId": "post_message_api_v1_conversations__cid__messages_post",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MessageIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Post Message",
    "tags": [
     "ai"
    ]
   }
  },
...
  "/api/v1/document-versions/{dv_id}/ai-extract": {
   "post": {
    "description": "Model-assisted extraction. Sends this document's extracted text to the `extract` route's provider.",
    "operationId": "ai_extract_api_v1_document_versions__dv_id__ai_extract_post",
    "parameters": [
     {
      "in": "path",
      "name": "dv_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Dv Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ai Extract",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/document-versions/{dv_id}/ai-policy": {
   "put": {
    "operationId": "set_policy_api_v1_document_versions__dv_id__ai_policy_put",
    "parameters": [
     {
      "in": "path",
      "name": "dv_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Dv Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProviderPolicy"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Set Policy",
    "tags": [
     "ai"
    ]
   }
  },
...
  "/api/v1/messages/{mid}/critique": {
   "post": {
    "operationId": "critique_api_v1_messages__mid__critique_post",
    "parameters": [
     {
      "in": "path",
      "name": "mid",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Mid",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Critique",
    "tags": [
     "ai"
    ]
   }
  },
  "/api/v1/programs": {
   "get": {
    "operationId": "list_programs_api_v1_programs_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ProgramOut"
         },
         "title": "Response List Programs Api V1 Programs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Programs",
    "tags": [
     "programs"
    ]
   },
   "post": {
    "operationId": "create_program_api_v1_programs_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProgramIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Program",
    "tags": [
     "ai",
     "programs"
    ]
   }
  },
...
  "/api/v1/reviews/bulk": {
   "post": {
    "description": "Apply one decision to many items. Each item is decided and audited individually; failures are reported.",
    "operationId": "bulk_decide_api_v1_reviews_bulk_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-cfs-user",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Cfs-User"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BulkDecision"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bulk Decide",
    "tags": [
     "ai",
     "reviews"
    ]
   }
  },
...
 }
}